+++ src/Avatar/index.ts
@@ -1,3 +1,2 @@
-export * from './PoweredByFooter';
 export { Avatar } from './Avatar';
 export type { AvatarProps } from './Avatar';